added input validation incl. prevent double-booking

+++ backend/src/controllers/bookings.js
@@ -21,6 +21,14 @@ const bookingController = Router()
 bookingController.get("/:id", auth(["admin", "trainer", "customer"]), async (req, res) => {
     const userID = req.params.id
 
+    if (!/^[1-9]\d*$/.test(userID)) {
+        res.status(400).json({
+            status: 400,
+            message: "Invalid request",
+        })
+        return
+    }
+
     Users.getByID(userID).then(user => {
         if (user.role == "trainer") {
             Classes.getByTrainerUserID(userID).then(classes => {
@@ -61,6 +69,21 @@ bookingController.get("/:id", auth(["admin", "trainer", "customer"]), async (req
 bookingController.post("/", auth(["admin", "trainer", "customer"]), async (req, res) => {
     const bookingData = req.body.booking
 
+    if (!/^[1-9]\d*$/.test(bookingData.userID)) {
+        res.status(400).json({
+            status: 400,
+            message: "Invalid request",
+        })
+        return
+    }
+    if (!/^[1-9]\d*$/.test(bookingData.classID)) {
+        res.status(400).json({
+            status: 400,
+            message: "Invalid request",
+        })
+        return
+    }
+
     // check if same class ID already exists in the user's booking table
     const existingBookings = await Bookings.getByUserID(bookingData.userID)
     if (existingBookings.some(booking => booking.class_id == bookingData.classID)) {
@@ -79,14 +102,14 @@ bookingController.post("/", auth(["admin", "trainer", "customer"]), async (req, 
         // breakdown time
         const [hours, minutes, seconds] = time.split(':').map(Number);
 
-        // Dateオブジェクトを作成します（年、月、日は任意で、時間だけが重要です）
-        // UTCを使用してDateオブジェクトを作成します（ローカルタイムゾーンに影響されないように）
+        // create new Date object, using UTC to avoid variation by different local time zones
+        // note: year, month, date are placeholders and not effecting the final result
         const date = new Date(Date.UTC(2000, 0, 1, hours, minutes, seconds));
 
-        // 分を追加します
+        // add minute
         date.setUTCMinutes(date.getUTCMinutes() + minutesToAdd);
 
-        // 新しい時間をHH:MM:SS形式で返します
+        // return new time in HH:MM:SS format
         return date.toISOString().substr(11, 8);
     }
 
@@ -104,46 +127,66 @@ bookingController.post("/", auth(["admin", "trainer", "customer"]), async (req, 
     })
     // console.log(existingBookingsOnSameDate)
 
-    // check if the time of the class being booked overlaps with the class times of the existing bookings
-    for (let i = 0; i < existingBookingsOnSameDate.length; i++) {
-        const existingBookingStartTime = existingBookingsOnSameDate[i].class_time
-        const existingBookingEndTime = addMinutesToTime(existingBookingStartTime, existingBookingsOnSameDate[i].activity_duration_minute)
-        console.log(i, existingBookingsOnSameDate[i], existingBookingStartTime, existingBookingEndTime)
+    if (existingBookingsOnSameDate == 0) {
+        // OK to book new class if:
+        // - no other class booked on the same day
 
-        if (classStartTime >= existingBookingEndTime || classEndTime <= existingBookingStartTime) {
-            // OK to book new class if:
-            // - its start time is later than existing booking's end time
-            // - its end time is earlier than existing booking's start time
+        console.log("no other class booked on the day, booking no problem!")
 
-            // console.log("booking no problem!")
-
-            Bookings.create(bookingData).then(booking => {
-                res.status(200).json({
-                    status: 200,
-                    message: "Created booking",
-                    booking: booking
-                })
-            }).catch(error => {
-                res.status(500).json({
-                    status: 500,
-                    message: "Failed to create booking:" + error,
-                })
+        Bookings.create(bookingData).then(booking => {
+            res.status(200).json({
+                status: 200,
+                message: "Created booking",
+                booking: booking
             })
-        } else if ((classStartTime >= existingBookingStartTime && classStartTime < existingBookingEndTime)
-            || (classEndTime > existingBookingStartTime && classEndTime <= existingBookingEndTime)
-            || (classStartTime < existingBookingStartTime && classEndTime > existingBookingEndTime)) {
-            // NG to book new class if:
-            // a) its start time is between existing booking's start time & end time
-            // b) its end time is between existing booking's start time & end time
-            // c) its start time is earlier than existing booking's start time
-            //  & its end time is later than existing booking's end time
-
-            // console.log("Failed to book: New class time overlaps with existing class bookings")
-
-            return res.status(400).json({
-                status: 400,
-                message: "Failed to book: New class time overlaps with existing class bookings"
+        }).catch(error => {
+            res.status(500).json({
+                status: 500,
+                message: "Failed to create booking:" + error,
             })
+        })
+    } else {
+        // check if the time of the class being booked overlaps with the class times of the existing bookings
+        for (let i = 0; i < existingBookingsOnSameDate.length; i++) {
+            const existingBookingStartTime = existingBookingsOnSameDate[i].class_time
+            const existingBookingEndTime = addMinutesToTime(existingBookingStartTime, existingBookingsOnSameDate[i].activity_duration_minute)
+            console.log(i, existingBookingsOnSameDate[i], existingBookingStartTime, existingBookingEndTime)
+
+            if (classStartTime >= existingBookingEndTime || classEndTime <= existingBookingStartTime) {
+                // OK to book new class if:
+                // - its start time is later than existing booking's end time
+                // - its end time is earlier than existing booking's start time
+
+                console.log("no class time overlap, booking no problem!")
+
+                Bookings.create(bookingData).then(booking => {
+                    res.status(200).json({
+                        status: 200,
+                        message: "Created booking",
+                        booking: booking
+                    })
+                }).catch(error => {
+                    res.status(500).json({
+                        status: 500,
+                        message: "Failed to create booking:" + error,
+                    })
+                })
+            } else if ((classStartTime >= existingBookingStartTime && classStartTime < existingBookingEndTime)
+                || (classEndTime > existingBookingStartTime && classEndTime <= existingBookingEndTime)
+                || (classStartTime < existingBookingStartTime && classEndTime > existingBookingEndTime)) {
+                // NG to book new class if:
+                // a) its start time is between existing booking's start time & end time
+                // b) its end time is between existing booking's start time & end time
+                // c) its start time is earlier than existing booking's start time
+                //  & its end time is later than existing booking's end time
+
+                // console.log("Failed to book: New class time overlaps with existing class bookings")
+
+                return res.status(400).json({
+                    status: 400,
+                    message: "Failed to book: New class time overlaps with existing class bookings"
+                })
+            }
         }
     }
 })
@@ -151,7 +194,14 @@ bookingController.post("/", auth(["admin", "trainer", "customer"]), async (req, 
 bookingController.delete("/:id", auth(["admin", "trainer", "customer"]), async (req, res) => {
     const bookingID = req.params.id
 
-    // TODO: Implement request validation
+    // Implement request validation
+    if (!/^[1-9]\d*$/.test(bookingID)) {
+        res.status(400).json({
+            status: 400,
+            message: "Invalid request",
+        })
+        return
+    }
 
     Bookings.deleteByID(bookingID).then(result => {
         res.status(200).json({
